modernized local media file handling with std::filesystem and std::ifstream

+++ common/arg.cpp
@@ -3304,9 +3304,9 @@ common_params_context common_params_parser_init(common_params & params, llama_ex
     ).set_examples({LLAMA_EXAMPLE_SERVER}).set_env("LLAMA_ARG_ALLOWED_LOCAL_MEDIA_PATH"));
     add_opt(common_arg(
         {"--local-media-max-size-mb"}, "N",
-        string_format("max size in mb for local media files (default: %d)", params.local_media_max_size_mb),
+        string_format("max size in mb for local media files (default: %lu)", params.local_media_max_size_mb),
         [](common_params & params, int value) {
-            params.local_media_max_size_mb = value;
+            params.local_media_max_size_mb = static_cast<size_t>(value);
         }
     ).set_examples({LLAMA_EXAMPLE_SERVER}).set_env("LLAMA_ARG_LOCAL_MEDIA_MAX_SIZE_MB"));
     add_opt(common_arg(
